Move 'EncodeAddress' and 'DecodeAddress' to 'Cardano.Wallet.Api.Types'
These instances really describe user-facing encoding that we only encounter at the API-level. So,
It felt wrong to have them down to the primitive 'AddressDiscovery' module
