+++ lib/core/src/Cardano/Wallet/Api/Types.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DeriveGeneric #-}
@@ -59,6 +60,10 @@ module Cardano.Wallet.Api.Types
     , ApiByronWalletMigrationInfo (..)
     , ByronWalletPostData (..)
 
+    -- * User-Facing Address Encoding/Decoding
+    , EncodeAddress (..)
+    , DecodeAddress (..)
+
     -- * Polymorphic Types
     , ApiT (..)
     , ApiMnemonicT (..)
@@ -74,8 +79,10 @@ import Cardano.Wallet.Primitive.AddressDerivation
     , PassphraseMaxLength (..)
     , PassphraseMinLength (..)
     )
-import Cardano.Wallet.Primitive.AddressDiscovery
-    ( DecodeAddress (..), EncodeAddress (..) )
+import Cardano.Wallet.Primitive.AddressDerivation.Byron
+    ( decodeLegacyAddress )
+import Cardano.Wallet.Primitive.AddressDerivation.Shelley
+    ( decodeShelleyAddress )
 import Cardano.Wallet.Primitive.AddressDiscovery.Sequential
     ( AddressPoolGap, getAddressPoolGap )
 import Cardano.Wallet.Primitive.Types
@@ -99,6 +106,8 @@ import Cardano.Wallet.Primitive.Types
     , WalletPassphraseInfo (..)
     , isValidCoin
     )
+import Codec.Binary.Bech32
+    ( dataPartFromBytes, dataPartToBytes, unsafeHumanReadablePartFromText )
 import Control.Applicative
     ( optional )
 import Control.Arrow
@@ -129,12 +138,18 @@ import Data.ByteArray.Encoding
     ( Base (Base16), convertFromBase )
 import Data.ByteString
     ( ByteString )
+import Data.ByteString.Base58
+    ( bitcoinAlphabet, decodeBase58, encodeBase58 )
 import Data.Either.Extra
     ( maybeToEither )
+import Data.Function
+    ( (&) )
 import Data.List.NonEmpty
     ( NonEmpty (..) )
 import Data.Map.Strict
     ( Map )
+import Data.Maybe
+    ( isJust )
 import Data.Proxy
     ( Proxy (..) )
 import Data.Quantity
@@ -162,6 +177,7 @@ import Servant.API
 import Web.HttpApiData
     ( FromHttpApiData (..), ToHttpApiData (..) )
 
+import qualified Codec.Binary.Bech32 as Bech32
 import qualified Data.Aeson as Aeson
 import qualified Data.Aeson.Types as Aeson
 import qualified Data.ByteString.Lazy as BL
@@ -781,3 +797,80 @@ taggedSumTypeOptions base opts = base
 
 eitherToParser :: Show s => Either s a -> Aeson.Parser a
 eitherToParser = either (fail . show) pure
+
+{-------------------------------------------------------------------------------
+                          User-Facing Address Encoding
+-------------------------------------------------------------------------------}
+
+-- | An abstract class to allow encoding of addresses depending on the target
+-- backend used.
+class EncodeAddress (n :: NetworkDiscriminant) where
+    encodeAddress :: Address -> Text
+
+-- | An abstract class to allow decoding of addresses depending on the target
+-- backend used.
+class DecodeAddress (n :: NetworkDiscriminant) where
+    decodeAddress :: Text -> Either TextDecodingError Address
+
+-- | Encode an 'Address' to a human-readable format. This produces two kinds of
+-- encodings:
+--
+-- - [Base58](https://en.wikipedia.org/wiki/Base58)
+--   for legacy / Byron addresses
+-- - [Bech32](https://github.com/bitcoin/bips/blob/master/bip-0173.mediawiki)
+--   for Shelley addresses
+--
+-- The right encoding is picked by looking at the raw 'Address' representation
+-- in order to figure out to which class the address belongs.
+instance EncodeAddress 'Mainnet where
+    encodeAddress = gEncodeAddress
+
+instance EncodeAddress 'Testnet where
+    encodeAddress = gEncodeAddress
+
+gEncodeAddress :: Address -> Text
+gEncodeAddress (Address bytes) =
+    if isJust (decodeLegacyAddress bytes) then base58 else bech32
+  where
+    base58 = T.decodeUtf8 $ encodeBase58 bitcoinAlphabet bytes
+    bech32 = Bech32.encodeLenient hrp (dataPartFromBytes bytes)
+      where hrp = unsafeHumanReadablePartFromText "addr"
+
+-- | Decode text string into an 'Address'. Jörmungandr recognizes two kind of
+-- addresses:
+--
+-- - Legacy / Byron addresses encoded as `Base58`
+-- - Shelley addresses, encoded as `Bech32`
+--
+-- See also 'EncodeAddress Jormungandr'
+instance DecodeAddress 'Mainnet where
+    decodeAddress = gDecodeAddress (decodeShelleyAddress @'Mainnet)
+
+instance DecodeAddress 'Testnet where
+    decodeAddress = gDecodeAddress (decodeShelleyAddress @'Testnet)
+
+gDecodeAddress
+    :: (ByteString -> Either TextDecodingError Address)
+    -> Text
+    -> Either TextDecodingError Address
+gDecodeAddress decodeShelley text =
+    case (tryBech32, tryBase58) of
+        (Just bytes, _) -> decodeShelley bytes
+        (_, Just bytes) -> decodeLegacyAddress bytes
+            & maybeToEither (TextDecodingError
+            "Unable to decode Address: neither Bech32-encoded nor a \
+            \valid Byron Address.")
+        (Nothing, Nothing) -> Left $ TextDecodingError
+            "Unable to decode Address: encoding is neither Bech32 nor \
+            \Base58."
+  where
+    -- | Attempt decoding a legacy 'Address' using a Base58 encoding.
+    tryBase58 :: Maybe ByteString
+    tryBase58 =
+        decodeBase58 bitcoinAlphabet (T.encodeUtf8 text)
+
+    -- | Attempt decoding an 'Address' using a Bech32 encoding.
+    tryBech32 :: Maybe ByteString
+    tryBech32 = do
+        (_, dp) <- either (const Nothing) Just (Bech32.decodeLenient text)
+        dataPartToBytes dp
