+++ lib/core/src/Cardano/Wallet/Primitive/AddressDiscovery.hs
@@ -5,7 +5,6 @@
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 
 -- |
@@ -26,8 +25,6 @@ module Cardano.Wallet.Primitive.AddressDiscovery
     , GenChange(..)
     , CompareDiscovery(..)
     , KnownAddresses(..)
-    , EncodeAddress(..)
-    , DecodeAddress(..)
     ) where
 
 import Prelude
@@ -35,32 +32,9 @@ import Prelude
 import Cardano.Crypto.Wallet
     ( XPrv )
 import Cardano.Wallet.Primitive.AddressDerivation
-    ( Depth (..), NetworkDiscriminant (..), Passphrase (..) )
-import Cardano.Wallet.Primitive.AddressDerivation.Byron
-    ( decodeLegacyAddress )
-import Cardano.Wallet.Primitive.AddressDerivation.Shelley
-    ( decodeShelleyAddress )
+    ( Depth (..), Passphrase (..) )
 import Cardano.Wallet.Primitive.Types
     ( Address (..) )
-import Codec.Binary.Bech32
-    ( dataPartFromBytes, dataPartToBytes, unsafeHumanReadablePartFromText )
-import Data.ByteString
-    ( ByteString )
-import Data.ByteString.Base58
-    ( bitcoinAlphabet, decodeBase58, encodeBase58 )
-import Data.Either.Extra
-    ( maybeToEither )
-import Data.Function
-    ( (&) )
-import Data.Maybe
-    ( isJust )
-import Data.Text
-    ( Text )
-import Data.Text.Class
-    ( TextDecodingError (..) )
-
-import qualified Codec.Binary.Bech32 as Bech32
-import qualified Data.Text.Encoding as T
 
 -- | This abstraction exists to give us the ability to keep the wallet business
 -- logic agnostic to the address derivation and discovery mechanisms.
@@ -150,80 +124,3 @@ class KnownAddresses s where
     knownAddresses
         :: s
         -> [Address]
-
--- | An abstract class to allow encoding of addresses depending on the target
--- backend used.
-class EncodeAddress (n :: NetworkDiscriminant) where
-    encodeAddress :: Address -> Text
-
--- | An abstract class to allow decoding of addresses depending on the target
--- backend used.
-class DecodeAddress (n :: NetworkDiscriminant) where
-    decodeAddress :: Text -> Either TextDecodingError Address
-
-{-------------------------------------------------------------------------------
-                                  Instances
--------------------------------------------------------------------------------}
-
--- | Encode an 'Address' to a human-readable format. This produces two kinds of
--- encodings:
---
--- - [Base58](https://en.wikipedia.org/wiki/Base58)
---   for legacy / Byron addresses
--- - [Bech32](https://github.com/bitcoin/bips/blob/master/bip-0173.mediawiki)
---   for Shelley addresses
---
--- The right encoding is picked by looking at the raw 'Address' representation
--- in order to figure out to which class the address belongs.
-instance EncodeAddress 'Mainnet where
-    encodeAddress = gEncodeAddress
-
-instance EncodeAddress 'Testnet where
-    encodeAddress = gEncodeAddress
-
-gEncodeAddress :: Address -> Text
-gEncodeAddress (Address bytes) =
-    if isJust (decodeLegacyAddress bytes) then base58 else bech32
-  where
-    base58 = T.decodeUtf8 $ encodeBase58 bitcoinAlphabet bytes
-    bech32 = Bech32.encodeLenient hrp (dataPartFromBytes bytes)
-      where hrp = unsafeHumanReadablePartFromText "addr"
-
--- | Decode text string into an 'Address'. Jörmungandr recognizes two kind of
--- addresses:
---
--- - Legacy / Byron addresses encoded as `Base58`
--- - Shelley addresses, encoded as `Bech32`
---
--- See also 'EncodeAddress Jormungandr'
-instance DecodeAddress 'Mainnet where
-    decodeAddress = gDecodeAddress (decodeShelleyAddress @'Mainnet)
-
-instance DecodeAddress 'Testnet where
-    decodeAddress = gDecodeAddress (decodeShelleyAddress @'Testnet)
-
-gDecodeAddress
-    :: (ByteString -> Either TextDecodingError Address)
-    -> Text
-    -> Either TextDecodingError Address
-gDecodeAddress decodeShelley text =
-    case (tryBech32, tryBase58) of
-        (Just bytes, _) -> decodeShelley bytes
-        (_, Just bytes) -> decodeLegacyAddress bytes
-            & maybeToEither (TextDecodingError
-            "Unable to decode Address: neither Bech32-encoded nor a \
-            \valid Byron Address.")
-        (Nothing, Nothing) -> Left $ TextDecodingError
-            "Unable to decode Address: encoding is neither Bech32 nor \
-            \Base58."
-  where
-    -- | Attempt decoding a legacy 'Address' using a Base58 encoding.
-    tryBase58 :: Maybe ByteString
-    tryBase58 =
-        decodeBase58 bitcoinAlphabet (T.encodeUtf8 text)
-
-    -- | Attempt decoding an 'Address' using a Bech32 encoding.
-    tryBech32 :: Maybe ByteString
-    tryBech32 = do
-        (_, dp) <- either (const Nothing) Just (Bech32.decodeLenient text)
-        dataPartToBytes dp
